requirements.txt
streamlit
langchain
langchain-groq
langchain-openai
langchain-pinecone
pinecone-client
python-dotenv